pyproject.toml
[project]
name = "score-recog"
version = "0.1.0"
description = ""
authors = [
    {name = "year",email = "sainknight87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "torchaudio (>=2.8.0,<3.0.0)",
    "transformers (>=4.56.2,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.0.0,<2.3.0)",
    "hf-xet (>=1.1.10,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "anthropic (>=0.69.0,<0.70.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "timm (>=1.0.20,<2.0.0)",
    "protobuf (>=6.33.0,<7.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.41.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu129"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch"}
torchvision = {source = "pytorch"}
torchaudio = {source = "pytorch"}
